Move from int page ids to long stored in embedding

+++ src/main/java/com/expleague/sensearch/donkey/plain/EmbeddingBuilder.java
@@ -33,7 +33,7 @@ public class EmbeddingBuilder {
 
   }
 
-  void add(int id, Vec v) {}
+  void add(long id, Vec v) {}
 
   void addAll(TIntObjectMap<Vec> vectors) {}
